Setup Controller
